rename send_map to hashmap
This makes the module much more discoverable, and is in line with the
'treemap' naming.
